refactor(frontend): group flights components into subfolders
