+++ apps/frontend/src/components/flights/create-site/CreateSiteModal.chromatic.stories.tsx
@@ -1,5 +1,5 @@
-import { FigureWrapper } from '../../../.storybook/FigureWrapper.tsx';
-import preview from '../../../.storybook/preview.tsx';
+import { FigureWrapper } from '../../../../.storybook/FigureWrapper.tsx';
+import preview from '../../../../.storybook/preview.tsx';
 import { Modal, SearchLoading } from './CreateSiteModal.stories.tsx';
 
 const meta = preview.meta({
