+++ apps/frontend/src/components/flights/strava-sync/StravaSyncModal.chromatic.stories.tsx
@@ -1,5 +1,5 @@
-import { FigureWrapper } from '../../../.storybook/FigureWrapper.tsx';
-import preview from '../../../.storybook/preview.tsx';
+import { FigureWrapper } from '../../../../.storybook/FigureWrapper.tsx';
+import preview from '../../../../.storybook/preview.tsx';
 import { Closed, Open } from './StravaSyncModal.stories.tsx';
 
 const meta = preview.meta({
