+++ apps/frontend/src/components/flights/table/mediaProgress.ts
