+++ apps/frontend/src/components/flights/video-export/VideoExportJobsPanel.tsx
@@ -15,9 +15,9 @@ import {
   useDeleteVideoExportJobRow,
   useResumeVideoExportJob,
   useVideoExportJobs,
-} from '../../hooks/flights/useVideoExportJobs';
-import { api } from '../../lib/api';
-import { useToast } from '../../hooks/useToast';
+} from '../../../hooks/flights/useVideoExportJobs';
+import { api } from '../../../lib/api';
+import { useToast } from '../../../hooks/useToast';
 
 const statusLabelFallbacks: Record<string, string> = {
   queued: 'En attente',
